refactor(routing): drop redundant HeadContent portal, extract cleanup hook
TanStack Router's docs render <HeadContent /> plainly for SPAs; it
manages document.head itself, so wrapping it in createPortal was
unnecessary. Extract the static-tag cleanup into useClearStaticTags so
its intent (clearing index.html's crawler-fallback tags once
HeadContent takes over) is named instead of inlined.
